requirements.txt
streamlit
requests
openai
python-dotenv
stripe
streamlit-authenticator
bcrypt
psycopg2-binary